requirements.txt
speedtest-cli
pymongo
